Fix conflict with Angular rendering cycle
When trying to use the pop-up menu within Angular, it breaks due to lifecycle conflicts with the Angular renderer. This just moves a call the the "parentElement" to a point in the Lit-Element lifecycle when that element is *certain* to be defined, rather than in the constructor, when Angular hasn't placed the element in the DOM yet.